# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export type NewsItemImportance = typeof NewsItemImportance[keyof typeof NewsItemImportance];


export const NewsItemImportance = {
  breaking: 'breaking',
  high: 'high',
  standard: 'standard',
} as const;

export interface NewsItem {
  id: string;
  title: string;
  content: string;
  source: string;
  publishedAt: string;
  relativeTime: string;
  categories: string[];
  relatedSymbols: string[];
  importance: NewsItemImportance;
  priceChange: number;
  impactScore: number;
  volumeChange: number;
}

export interface MarketAsset {
  symbol: string;
  name: string;
  price: number;
  change24h: number;
  volume24h: string;
  marketCap: string;
  high24h: number;
  low24h: number;
  sparkline: number[];
  accent: string;
}

export type SentimentLabel = typeof SentimentLabel[keyof typeof SentimentLabel];


export const SentimentLabel = {
  GREED: 'GREED',
  FEAR: 'FEAR',
  NEUTRAL: 'NEUTRAL',
} as const;

export interface Sentiment {
  score: number;
  label: SentimentLabel;
  change: number;
}

export interface MarketOverview {
  assets: MarketAsset[];
  sentiment: Sentiment;
  totalMarketCap: string;
  totalVolume: string;
  btcDominance: string;
  updatedAt: string;
}

export interface MarketMovers {
  gainers: MarketAsset[];
  losers: MarketAsset[];
}

export interface NewsImpact {
  newsId: string;
  score: number;
  priceChange: number;
  volumeChange: number;
}

/**
 * The requested market item was not found
 */
export type NotFoundResponse = ErrorResponse;