# lib/api-zod/src/generated/types/marketMovers.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MarketAsset } from './marketAsset';

export interface MarketMovers {
  gainers: MarketAsset[];
  losers: MarketAsset[];
}