# lib/api-zod/src/generated/types/marketOverview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MarketAsset } from './marketAsset';
import type { Sentiment } from './sentiment';

export interface MarketOverview {
  assets: MarketAsset[];
  sentiment: Sentiment;
  totalMarketCap: string;
  totalVolume: string;
  btcDominance: string;
  updatedAt: string;
}