# lib/api-zod/src/generated/types/newsItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NewsItemImportance } from './newsItemImportance';

export interface NewsItem {
  id: string;
  title: string;
  content: string;
  source: string;
  publishedAt: Date;
  relativeTime: string;
  categories: string[];
  relatedSymbols: string[];
  importance: NewsItemImportance;
  priceChange: number;
  impactScore: number;
  volumeChange: number;
}