# lib/api-zod/src/generated/types/sentiment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SentimentLabel } from './sentimentLabel';

export interface Sentiment {
  score: number;
  label: SentimentLabel;
  change: number;
}